chat-agent-store seems to be working, gonna try full scale implementation
