+++ elements/chat-agent/lib/chat-interface.js
@@ -3,6 +3,7 @@
  * @license Apache-2.0, see License.md for full text.
  */
 import { ChatAgentModalStore } from "../chat-agent.js";
+import { ChatStore } from "./chat-agent-store.js";
 import { DDD } from "@haxtheweb/d-d-d/d-d-d.js";
 import { autorun, toJS, } from "mobx";
 import { html, css } from "lit";
