+++ elements/chat-agent/src/chat-agent.js
@@ -591,291 +591,3 @@ globalThis.ChatAgentModal.requestAvailability = () => {
 // ! Chat Agent Store
 export const ChatAgentModalStore =
   globalThis.ChatAgentModal.requestAvailability();
-
-// TODO can read from store, but cannot write to store. Also encountering "ReferenceError: can't access lexical declaration 'chatAgentStore' before initialization"
-class ChatAgentStore {
-
-  constructor() {
-    this.buttonIcon = "hax:wizard-hat";
-    this.chatLog = [];
-    this.context = "phys211";
-    this.darkMode = null;
-    this.dataCollectionEnabled = true;
-    this.developerModeEnabled = false;
-    this.editMode = null;
-    this.engine = "alfred";
-    this.isFullView = false;
-    this.isInterfaceHidden = true; // TODO switch to false for testing when I know this store works
-    this.isLoading = null;
-    this.merlinIndex = 0;
-    this.messageIndex = 0;
-    this.userIndex = 0;
-
-    makeObservable(this, {
-      buttonIcon: observable,
-      chatLog: observable,
-      context: observable,
-      darkMode: observable,
-      dataCollectionEnabled: observable,
-      developerModeEnabled: observable,
-      editMode: observable,
-      engine: observable,
-      isFullView: observable,
-      isInterfaceHidden: observable,
-      isLoading: observable,
-      merlinIndex: observable,
-      messageIndex: observable,
-      userIndex: observable,
-    });
-  
-    autorun(() => {
-      const buttonIcon = toJS(this.buttonIcon);
-      const chatLog = toJS(this.chatLog);
-      const context = toJS(this.context);
-      const darkMode = toJS(store.darkMode);
-      const dataCollectionEnabled = toJS(this.dataCollectionEnabled);
-      const developerModeEnabled = toJS(this.developerModeEnabled);
-      const editMode = toJS(this.editMode);
-      const engine = toJS(this.engine);
-      const isFullView = toJS(this.isFullView);
-      const isInterfaceHidden = toJS(this.isInterfaceHidden);
-      const isLoading = toJS(this.isLoading);
-      const merlinIndex = toJS(this.merlinIndex);
-      const messageIndex = toJS(this.messageIndex);
-      const userIndex = toJS(this.userIndex);
-    });
-  
-    if (isLoading) {
-      this.buttonIcon = "hax:loading";
-    } else {
-      this.buttonIcon = "hax:wizard-hat";
-    }
-  }
-
-  /**
-   * @description writes message to chatLog
-   * @param {string} author - the author of the message
-   * @param {string} message - the written or suggested prompt
-   */
-  handleMessage(author, message) {
-    this.devStatement(`Writing message "${message}" by ${author} to chatLog.`, `info`);
-
-    let authorIndex;
-
-    this.messageIndex++;
-
-    switch (author) {
-      case "merlin":
-        this.merlinIndex++;
-        authorIndex = this.merlinIndex;
-        break;
-      case this.userName:
-        this.userIndex++;
-        authorIndex = this.userIndex;
-        break;
-    }
-
-    let date = new Date();
-
-    const chatLogObject = {
-      messageID: this.messageIndex,
-      author: author,
-      message: message,
-      authorMessageIndex: authorIndex,
-      timestamp: date.toString().replace(/\s/g, "-"),
-    };
-
-    this.chatLog.push(chatLogObject);
-
-    if (author === this.userName) {
-      this.handleInteraction(message);
-    }
-  }
-
-  /**
-   * @description sends prompt to AI engine specified
-   * @param {string} prompt - the written or suggested prompt
-   */
-  handleInteraction(prompt) {
-    this.devStatement(`Prompt sent to: ${this.engine}. Prompt sent: ${prompt}`, `info`)
-    this.currentSuggestions = [];
-
-    switch (prompt) {
-      // Offline messages, do not request response from backend AI
-
-      // Tutorial messages
-      case "Who are you?":
-        this.currentSuggestions = [
-          {
-            suggestion: "What can you do for me?",
-            type: "help",
-          },
-          {
-            suggestion: "How do I use you?",
-            type: "help",
-          },
-        ];
-        this.handleMessage(
-          "merlin",
-          "I am Merlin. I was created for use within HAX websites as an assistant to help you with your questions. How may I help you today?",
-        );
-        break;
-      case "What can you do for me?":
-        this.currentSuggestions = [
-          {
-            suggestion: "Who are you?",
-            type: "hax",
-          },
-          {
-            suggestion: "How do I use you?",
-            type: "help",
-          },
-        ];
-        this.handleMessage(
-          "merlin",
-          "I can answer questions and chat with you about information relevant to the website you are navigating. How can I help you?",
-        );
-        break;
-      case "How do I use you?":
-        this.currentSuggestions = [
-          {
-            suggestion: "Who are you?",
-            type: "hax",
-          },
-          {
-            suggestion: "What can you do for me?",
-            type: "help",
-          },
-        ];
-        this.handleMessage(
-          "merlin",
-          "I support numerous functions. You can ask me questions, as well as download our chat log and reset our chat. You can start asking me questions by clicking on one of the suggested prompts, or by typing a prompt in the input box below and pressing the send button or pressing the enter key on your keyboard. Here are some of the keyboard controls you can utilize: \n 1. Tab Key - Navigates you through the numerous usable buttons. \n 2. Enter Key (in text area) - Will submit the prompt you wrote. \n 3. Enter key (When focusing on a button) - Will act in the same way as clicking the button. \n 4. Up & Down Arrow Keys (in text area) - will navigate you through previously sent prompts so you can send them again.",
-        );
-        break;
-
-      // Network error messages
-      case "Why can't you connect?":
-        this.currentSuggestions = [
-          {
-            suggestion: "How do I fix this connection issue?",
-            type: "network",
-          },
-        ];
-        this.handleMessage(
-          "merlin",
-          "I am either unable to connect to the internet, or a service I connect to is not available, meaning I cannot research how to respond to your prompt.",
-        );
-        break;
-      case "How do I fix this connection issue?":
-        this.currentSuggestions = [
-          {
-            suggestion: "Why can't you connect?",
-            type: "network",
-          },
-        ];
-        this.handleMessage(
-          "merlin",
-          "Please ensure you are connected to the internet. I cannot respond to (most of) your questions if you are not connected to the internet. If you are connected, it is likely one of my connected services is having an issue, I will try to fix that and be back to help you soon.",
-        );
-        break;
-
-      // Online messages, do request response from backend AI
-      default:
-        var base = "";
-
-        if (globalThis.document.querySelector("base")) {
-          base = globalThis.document.querySelector("base").href;
-        }
-
-        // TODO: Add support for data collection toggle
-        const params = {
-          site: {
-            file: "https://haxtheweb.org/site.json",
-          },
-          type: "site",
-          question: prompt,
-          engine: this.engine,
-          context: this.context,
-        };
-
-        this.isLoading = true;
-
-        MicroFrontendRegistry.call("@haxcms/aiChat", params)
-          .then((d) => {
-            if (d.status == 200) {
-              this.answers = [d.data.answers];
-              this.question = d.data.question;
-              this.currentSuggestions = []; // TODO add support for AI based suggestions
-            }
-
-            this.isLoading = false;
-            this.handleMessage("merlin", d.data.answers);
-          })
-          .catch((error) => {
-            this.isLoading = false;
-            this.currentSuggestions = [
-              {
-                suggestion: "Why can't you connect?",
-                type: "network",
-              },
-              {
-                suggestion: "How do I fix this connection issue?",
-                type: "network",
-              },
-            ];
-            this.handleMessage(
-              "merlin",
-              "I'm sorry, I'm having trouble connecting right now. Please try again soon. If you'd like to learn more, please click on one of the suggested prompts.",
-            );
-            console.error(error);
-          });
-    }
-  }
-
-  /**
-   * @description downloads the chat log as the specified file type
-   * @param {string} fileType - the file type to download
-   */
-  handleDownload(fileType) {
-    this.devStatement(`Downloading chatlog as ${fileType}.`, 'info');
-
-    if (this.chatLog.length !== 0) {
-      const LOG = JSON.stringify(this.chatLog, undefined, 2);
-      let date = new Date();
-      const FILE_NAME = `${this.userName}-chat-log-${date.toString().replace(/\s/g, "-")}.${fileType}`;
-
-      let download = document.createElement("a");
-      download.setAttribute(
-        "href",
-        "data:text/plain;charset=utf-8," + encodeURIComponent(LOG),
-      );
-      download.setAttribute("download", FILE_NAME);
-      download.click();
-      download.remove();
-    }
-  }
-
-  devStatement(statement, type) {
-    if (this.developerModeEnabled) {
-      switch (type) {
-        case "log":
-          console.log(`CHAT-AGENT-DEV-MODE: ${statement}`);
-          break;
-        case "info":
-          console.info(`CHAT-AGENT-DEV-MODE: ${statement}`);
-          break;
-        case "warn":
-          console.warn(`CHAT-AGENT-DEV-MODE: ${statement}`);
-          break;
-        case "error":
-          console.error(`CHAT-AGENT-DEV-MODE: ${statement}`);
-          break;
-
-        default:
-          console.error("No devStatement type specified");
-      }
-    }
-  }
-}
-
-export const ChatStore = globalthis.ChatAgentStore.requestAvailability();
